User canBeEliminated method and view

+++ resources/views/users/index.blade.php
@@ -54,11 +54,13 @@
                                 @endif
                                 @if(Auth::user()->can('users/destroy'))
                                     <td>
-                                        {!! Form::open(['method' => 'DELETE', 'action' => ['UsersController@destroy', $user->id]]) !!}
-                                            <button class="btn-link" type="submit">
-                                                <i class="fa fa-trash"></i>
-                                             </button>
-                                        {!! Form::close() !!}
+                                        @if($user->canBeEliminated())
+                                            {!! Form::open(['method' => 'DELETE', 'action' => ['UsersController@destroy', $user->id]]) !!}
+                                                <button class="btn-link" type="submit">
+                                                    <i class="fa fa-trash"></i>
+                                                 </button>
+                                            {!! Form::close() !!}
+                                        @endif
                                     </td>
                                 @endif
                             </tr>
